+++ project/src/components/UI/footer/Footer.tsx
@@ -1,0 +1,9 @@
+export default function Footer() {
+  return (
+    <footer className="footer">
+      <a className="footer__logo-link" href="/">
+        <img className="footer__logo" src="img/logo.svg" alt="6 cities logo" width="64" height="33" />
+      </a>
+    </footer>
+  );
+}
