+++ project/src/mocks/reviews.ts
@@ -1,6 +1,6 @@
 import { Favorites } from '../types/hotel';
 
-export const favorites:Favorites = [
+export const favorites:Favorites[] = [
   {
     city: 'Amsterdam',
     list: [
@@ -16,42 +16,6 @@ export const favorites:Favorites = [
         name: 'Beautiful &amp; luxurious apartment at great location',
         type: 'Apartment',
       },
-      {
-        id: 2,
-        premium: false,
-        img: {
-          src: 'img/room.jpg',
-          alt: 'Place image',
-        },
-        price: '80',
-        rating: 4,
-        name: 'Wood and stone place',
-        type: 'Private room',
-      },
-      {
-        id: 3,
-        premium: false,
-        img: {
-          src: 'img/apartment-02.jpg',
-          alt: 'Place image',
-        },
-        price: '132',
-        rating: 4,
-        name: 'Canal View Prinsengracht',
-        type: 'Apartment',
-      },
-      {
-        id: 4,
-        premium: true,
-        img: {
-          src: 'img/apartment-03.jpg',
-          alt: 'Place image',
-        },
-        price: '180',
-        rating: 5,
-        name: 'Nice, cozy, warm big bed apartment',
-        type: 'Apartment',
-      },
     ],
   },
   {
@@ -108,7 +72,7 @@ export const favorites:Favorites = [
     ],
   },
   {
-    city: 'Paris',
+    city: 'Hamburg',
     list: [
       {
         id: 1,
@@ -144,24 +108,12 @@ export const favorites:Favorites = [
         price: '132',
         rating: 4,
         name: 'Canal View Prinsengracht',
-        type: 'Apartment',
-      },
-      {
-        id: 4,
-        premium: true,
-        img: {
-          src: 'img/apartment-03.jpg',
-          alt: 'Place image',
-        },
-        price: '180',
-        rating: 5,
-        name: 'Nice, cozy, warm big bed apartment',
         type: 'Apartment',
       },
     ],
   },
   {
-    city: 'Paris',
+    city: 'Cologne',
     list: [
       {
         id: 1,
@@ -186,30 +138,6 @@ export const favorites:Favorites = [
         rating: 4,
         name: 'Wood and stone place',
         type: 'Private room',
-      },
-      {
-        id: 3,
-        premium: false,
-        img: {
-          src: 'img/apartment-02.jpg',
-          alt: 'Place image',
-        },
-        price: '132',
-        rating: 4,
-        name: 'Canal View Prinsengracht',
-        type: 'Apartment',
-      },
-      {
-        id: 4,
-        premium: true,
-        img: {
-          src: 'img/apartment-03.jpg',
-          alt: 'Place image',
-        },
-        price: '180',
-        rating: 5,
-        name: 'Nice, cozy, warm big bed apartment',
-        type: 'Apartment',
       },
     ],
   },
