+++ project/src/pages/FavoritePage.tsx
@@ -1,61 +1,49 @@
-import CardFavorites from '../components/CardFavorites';
+import CardFavorites from '../components/UI/Сard/CardFavorites';
 import { Favorites } from '../types/hotel';
+import Footer from '../components/UI/footer/Footer';
+import Header from '../components/UI/header/Header';
 
 type FavoritePageProps = {
-  favorites: Favorites;
+  favorites: Favorites[];
 }
 
 export default function FavoritePage({ favorites }: FavoritePageProps) {
-  // eslint-disable-next-line no-console
-  console.log(favorites);
+  if (!favorites.length) {
+    return (
+      <div className="page page--favorites-empty">
+        <Header />
+        <main className="page__main page__main--favorites page__main--favorites-empty">
+          <div className="page__favorites-container container">
+            <section className="favorites favorites--empty">
+              <h1 className="visually-hidden">Favorites (empty)</h1>
+              <div className="favorites__status-wrapper">
+                <b className="favorites__status">Nothing yet saved.</b>
+                <p className="favorites__status-description">Save properties to narrow down search or plan your future trips.</p>
+              </div>
+            </section>
+          </div>
+        </main>
+        <Footer />
+      </div>
+    );
+  }
+
   return (
     <div className="page">
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <a className="header__logo-link" href="main.html">
-                <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41" />
-              </a>
-            </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <a className="header__nav-link header__nav-link--profile" href="/">
-                    <div className="header__avatar-wrapper user__avatar-wrapper">
-                    </div>
-                    <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
-                  </a>
-                </li>
-                <li className="header__nav-item">
-                  <a className="header__nav-link" href="/">
-                    <span className="header__signout">Sign out</span>
-                  </a>
-                </li>
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </header>
-
+      <Header />
       <main className="page__main page__main--favorites">
         <div className="page__favorites-container container">
           <section className="favorites">
             <h1 className="favorites__title">Saved listing</h1>
             <ul className="favorites__list">
-              {favorites.map((favorite, index) =>
-                // eslint-disable-next-line react/jsx-key
-                <CardFavorites {...favorite} />,
+              {favorites.map((favorite) =>
+                <CardFavorites {...favorite} key={favorite.city} />,
               )}
             </ul>
           </section>
         </div>
       </main>
-      <footer className="footer container">
-        <a className="footer__logo-link" href="main.html">
-          <img className="footer__logo" src="img/logo.svg" alt="6 cities logo" width="64" height="33" />
-        </a>
-      </footer>
+      <Footer />
     </div>
   );
 }
